Revert "Hash Entry size optimisation real"
This reverts commit fcf36d98ac838324adab66849db6bf91e44ae865.

+++ src/defs.h
@@ -93,16 +93,15 @@ typedef struct {
 	int count;
 } S_MOVELIST;
 
-// Move flag for hash entry
 enum { HFNONE, HFALPHA, HFBETA, HFEXACT };
 
 // 18 bytes (will be padded to 24)
 typedef struct {
 	U64 posKey;
 	int move;
-	int16_t score;
+	int score;
 	uint8_t depth;
-	uint8_t flags;
+	int flags;
 	uint16_t age; // indicates how new an entry is
 } S_HASHENTRY;
